+++ app/Record.php
@@ -25,6 +25,8 @@ use Illuminate\Database\Eloquent\Model;
  *
  * Represents a DNS entry on a specified zone.
  *
+ * Valid records types are defined on DNSHelper Class.
+ *
  * @property integer $id        The object unique id.
  * @property string  $name      The name of the record.
  * @property integer $ttl       The custom TTL value for this record.
@@ -59,7 +61,7 @@ class Record extends Model
         'ttl'      => 'integer',
         'type'     => 'string',
         'priority' => 'integer',
-        'data'     => 'string'
+        'data'     => 'string',
     ];
     /**
      * The attributes that should be casted to null if is empty.
@@ -76,25 +78,6 @@ class Record extends Model
      * @var array
      */
     protected $touches = ['zone'];
-
-    /**
-     * Return an array with ALL valid Record types.
-     *
-     * @return array
-     */
-    public static function getAllValidRecordTypes() : array
-    {
-        return [
-            'A'     => trans('record/model.types_mapper.A'),
-            'AAAA'  => trans('record/model.types_mapper.AAAA'),
-            'CNAME' => trans('record/model.types_mapper.CNAME'),
-            'MX'    => trans('record/model.types_mapper.MX'),
-            'NS'    => trans('record/model.types_mapper.NS'),
-            'PTR'   => trans('record/model.types_mapper.PTR'),
-            'SRV'   => trans('record/model.types_mapper.SRV'),
-            'TXT'   => trans('record/model.types_mapper.TXT'),
-        ];
-    }
 
     /**
      * Set the Record's type uppercase.
@@ -117,16 +100,6 @@ class Record extends Model
     }
 
     /**
-     * Set the Record's data lowercase.
-     *
-     * @param  string $value
-     */
-    public function setDataAttribute(string $value)
-    {
-        $this->attributes['data'] = strtolower($value);
-    }
-
-    /**
      * A record belongs to a zone.
      *
      * This is a one-to-one relationship.
@@ -142,7 +115,7 @@ class Record extends Model
      * @return string
      * @codeCoverageIgnore
      */
-    public function getResourceRecord() : string
+    public function getResourceRecord(): string
     {
         if ($this->ttl) {
             return sprintf("%-40s %d\tIN\t%s\t%s", $this->name, $this->ttl, $this->type, $this->data);