+++ CHANGELOG.md
@@ -4,11 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- Add support to [NAPRT DNS record type](https://en.wikipedia.org/wiki/NAPTR_record). 
+
+### Changed
+- Move valid record types definition and `validateRecordType()` method to `DNSHelper` class.
+
+### Fixed
+- Fix an error in Record model. Data attribute was lowercase, but this data field may contain uppercase and lowercase.
 
 ## 0.8.1 - 2018-05-20
 ### Fixed
 - Fix AdminLTE paths in order to be compatible with v2.3.11 (admin-lte)
-
 
 ## 0.8.0 - 2018-05-20
 ### Changed